int to Long. Add constructor instead of @Autowired. Update readme file.

+++ src/main/java/com/example/springBootDemo/model/Employee.java
@@ -16,7 +16,6 @@ public class Employee {
 
     @Id
     @GeneratedValue
-    private int id;
+    private Long id;
     private String name;
-
 }
